requirements.txt
flask==0.10.1
flask-sqlalchemy==2.0
pymysql==0.6.6
